[#166] refactor: 캐릭터 조회 폴더구조 변경
