+++ src/app/(route)/(main)/character/page.tsx
@@ -1,19 +1,16 @@
-import ChracterData from '@/app/_components/character/ChracterData';
-import Loading from '@/app/_components/common/Loading';
-import RoundedContainer from '@/app/_components/layout/RoundedContainer';
-import PreviewSearchBar from '@/app/_components/preview/previewSearchMenuBar/previewSearchBar/PreviewSearchBar';
-import { Suspense } from 'react';
+import CharacterSearchInput from '@/app/_components/CharacterSearchInput';
+import BasicContainer from '@/app/_components/layout/BasicContainer';
+import { ChracterInformation } from '@/app/_features/character';
 
 const Page = () => {
   return (
-    <Suspense fallback={<Loading />}>
-      <div className="dark flex flex-1 flex-col gap-2 p-2">
-        <RoundedContainer className="flex h-14 items-center justify-center">
-          <PreviewSearchBar className="w-full max-w-72" routeName="character" />
-        </RoundedContainer>
-        <ChracterData />
-      </div>
-    </Suspense>
+    <BasicContainer className="flex-1 gap-2">
+      <CharacterSearchInput
+        className="mx-auto h-8 w-full max-w-72"
+        routeName="character"
+      />
+      <ChracterInformation />
+    </BasicContainer>
   );
 };
