+++ src/app/_components/CharacterSearchInput.tsx
@@ -4,7 +4,7 @@ import Search from '@/app/_components/common/search/Search';
 import { useRefFocus } from '@/app/_hooks/useRefFocus/useRefFocus';
 import { useSearchHandler } from '@/app/_hooks/useSearchHandler/useSearchHandler';
 
-const PreviewSearchBar = ({
+const CharacterSearchInput = ({
   className = '',
   routeName,
 }: {
@@ -23,4 +23,4 @@ const PreviewSearchBar = ({
   );
 };
 
-export default PreviewSearchBar;
+export default CharacterSearchInput;
