+++ src/app/_features/character/components/equipment/grid/CachItem.tsx
@@ -1,0 +1,34 @@
+import ItemPrviewImage from '@/app/_components/iteminfo/info/ItemPrviewImage';
+import CachColors from './CachColors';
+import { CachItemsType } from '@/app/_type/equipmentType';
+
+const CachItem = ({ ...item }: CachItemsType) => {
+  return (
+    <div className="flex h-full w-full flex-col items-center rounded-md border border-border">
+      <div className="flex h-full w-full items-center justify-center">
+        <div className="flex h-full w-full flex-col justify-center gap-1 p-2 px-1 hover:bg-muted/70">
+          <div className="flex w-full items-center gap-1">
+            <div className="hidden lg:block">
+              <ItemPrviewImage
+                materials={item.item_name}
+                slot={item.item_equipment_slot_name}
+              />
+            </div>
+            <div
+              className="truncate text-center text-xs"
+              title={item.item_name}
+            >
+              {item.item_name}
+            </div>
+          </div>
+          <CachColors
+            isWearing={!item.item_name.includes('미착용')}
+            colors={item.item_option.cash_item_color}
+          />
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default CachItem;
