+++ src/app/_features/character/components/equipment/grid/index.ts
@@ -1,0 +1,2 @@
+export { default as EquipmentGrid } from './EquipmentGrid';
+export { default as CachGrid } from './CachGrid';
