+++ src/app/_features/character/components/equipment/index.ts
@@ -1,0 +1,1 @@
+export { default as CharacterEquipment } from './CharacterEquipment';
