+++ src/app/_features/character/components/information/index.ts
@@ -1,0 +1,1 @@
+export { default as CharacterBasicInfo } from './CharacterBasicInfo';
