+++ src/app/_features/character/components/skills/index.ts
@@ -1,0 +1,1 @@
+export { default as SkillAwakeningTable } from './SkillAwakeningTable';
