+++ src/app/_features/character/constant/index.ts
@@ -1,0 +1,23 @@
+export const ITEM_SLOT = [
+  ['Earring', 'Head', 'Necklace'],
+  ['Right Hand', 'Upper', 'Left Hand'],
+  [null, 'Lower', 'Hand'],
+  ['Belt', 'Leg', 'Charm'],
+  ['Right Finger', 'Artifact', 'Left Finger'],
+  ['Right Wrist', 'Rhod', 'Left Wrist'],
+];
+
+export const AVATAR_SLOT = [
+  [null, 'Avatar_Helm', 'Avatar_Rear'],
+  ['Avatar_Weapon', 'Avatar_Tunic', null],
+  // ["Avatar_Weapon", "Avatar_Tunic", "(Unknown)"],
+  [null, 'Avatar_Pants', 'Avatar_Gloves'],
+  ['Avatar_Tail', 'Avatar_Boots', null],
+];
+
+export const CACH_SLOT = [
+  ['Right Epaulet', 'Hair', 'Left Epaulet'],
+  ['FacePainting', 'Lens', 'Scar'],
+  ['Inner Armor', 'MakeUp', null],
+  ['Body Shape', 'BodyPainting', 'Badge'],
+];
