+++ src/app/_features/character/index.ts
@@ -1,0 +1,1 @@
+export { default as ChracterInformation } from './components/ChracterInformation';
