+++ src/app/_features/character/utils/getSkillDescription.ts
@@ -1,0 +1,7 @@
+import { skillDescription } from '@/app/_constant/character/skill/skill';
+
+export const getSkillDescription = (skillName: string, jobName: string) => {
+  const skills = skillDescription[jobName] || [];
+  if (skills.length === 0) return '';
+  return skills?.find((s) => s.name === skillName.trim())?.effect;
+};
