+++ src/app/_features/iteminfo/constant/index.ts
